Split the chance cube's tuning from its Discord presentation
The rules and the emoji were one file, so the game could only ever be played
in an embed. Every face now names itself with an abstract id -- the same
`kind` / `kind:side` shape `faceKey` already produced -- and what an id looks
like moved to data/discord/cube_emoji.js, alongside the pacing and layout
config that only exists because a frame is a message edit.

data/challenge/cube.js becomes a shim that dresses the pure tuning back into
the shape the embed was written against, so nothing downstream changes yet.
There is still exactly one source of truth: the shim redeclares no values,
it only re-attaches glyphs.

Ben's wings carry their handedness as ids rather than emoji, so `mirrored`
keeps working -- reflected, his left third still has to come back as his
right or the glass draws him inside out.

Verified against the committed original: every key it declared is reproduced
exactly, with `id` the only addition, and 20k full-rack rolls resolve with no
undefined glyphs.
